pyproject.toml
[project]
name = "csbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
]
